docs/examples/stellar_optimization/requirements.txt
# Additional Python dependencies for this workload
# (Base constellaration dependencies come from the base image)

# Cloud storage
google-cloud-storage>=2.10.0

# Ray for distributed computing
ray[default]>=2.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
